🎨 UserCard에서 데이터 받아오는 부분 Listpage로 변경
